grpc_server/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
joblib==1.4.2
numpy==2.1.1
pandas==2.2.2
protobuf==5.28.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
